+++ src/vcsUiApp.js
@@ -45,6 +45,7 @@ import { createVueI18n, setupI18n } from './vuePlugins/i18n.js';
 import { callbackClassRegistry } from './callback/vcsCallback.js';
 import createSiteConfig from './siteConfig.js';
 import { createVcsVuetify } from './vuePlugins/vuetify.js';
+import createObliqueFallbackWarnings from './obliqueFallbackWarnings.js';
 
 /**
  * @typedef {import("@vcmap/core").VcsModuleConfig & {
@@ -374,6 +375,8 @@ class VcsUiApp extends VcsApp {
     this.mounted = new VcsEvent();
 
     this._destroySiteConfig = createSiteConfig(this._uiConfig);
+
+    this._destroyObliqueFallback = createObliqueFallbackWarnings(this);
   }
 
   /**
@@ -753,6 +756,7 @@ class VcsUiApp extends VcsApp {
     this._featureInfo.destroy();
     this._uiConfig.destroy();
     this._destroySiteConfig();
+    this._destroyObliqueFallback();
     super.destroy();
   }
 }