+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 5.3.6
+
+- Added export `LegendTypes` and `StyleRowType` enums and typedefs from `legendHelper.js`
+
+# 5.3.5
+
+- Oblique fallback warnings have been re-instated as notifications.
+
+# 5.3.4
+
+- Update @vcmap/core to 5.3.2 to fix a reading legacy oblique datasets
+
+# 5.3.3
+
+- Updates `splashScreen` to use the filled variant of the button.
+- Fixes bug in `createSupportedMapMappingFunction` where `mapCollection.activeMap === null` caused an error
+- Update @vcmap/core to 5.3.1 to fix a Declarative Style Icon rendering Problem in 2D
+
 # 5.3.2
 
 - Fixes Bug in `splashScreen`